Fix typo and move to "ExternalResources" class with tests
